Add pep-249 dbapi module
See [1] for details. The main motivation for the module creation was
the integration Django with Tarantool database through django-tarantool
database backend [2] which requires dbapi connector for the database.

The most of the optional extensions and methods were ignored because
Django does not require them. Anyway, feel free to suggest its
implementation as needed.

Interactive transactions are not currently supported by Tarantool and
theirs implementation will be added in the connector when the feature
is stable in Tarantool itself.

[1] https://www.python.org/dev/peps/pep-0249/
[2] https://github.com/artembo/django-tarantool

Co-authored-by: Denis Ignatenko <denis.ignatenko@gmail.com>